Update libraries by Arduino UI

+++ libraries/WiFiNINA/src/utility/spi_drv.cpp
@@ -30,7 +30,7 @@
 // yes, so use the existing VidorFPGA include
 #include <VidorFPGA.h>
 #else
-// otherwise, fallback to VidorPeripherals and it's bistream
+// otherwise, fallback to VidorPeripherals and it's bitstream
 #include <VidorPeripherals.h>
 #endif
 
@@ -451,7 +451,7 @@ int SpiDrv::waitResponse(uint8_t cmd, uint8_t* numParamRead, uint8_t** params, u
 
 void SpiDrv::sendParamNoLen(uint8_t* param, size_t param_len, uint8_t lastParam)
 {
-    int i = 0;
+    size_t i = 0;
     // Send Spi paramLen
     sendParamLen8(0);
